Very basic comment CAPTCHA
This randomly generates a simple math problem that the user must answer
correctly in order to post a comment. Problems include the use of global
variables, insufficient frequency number of randomization, and a lack of
feedback when the user screws up or is a bot. I should work on those at
some point.